Llms: separate client transport functions
