+++ src/modules/aifn/autotitle/autoTitle.ts
@@ -1,4 +1,4 @@
-import { callChatGenerate } from '~/modules/llms/llm.client';
+import { callChatGenerate } from '~/modules/llms/transports/chatGenerate';
 import { useModelsStore } from '~/modules/llms/store-llms';
 
 import { useChatStore } from '~/common/state/store-chats';
