+++ src/modules/aifn/flatten/flatten.ts
@@ -1,5 +1,5 @@
 import { DLLMId } from '~/modules/llms/store-llms';
-import { callChatGenerate } from '~/modules/llms/llm.client';
+import { callChatGenerate } from '~/modules/llms/transports/chatGenerate';
 
 import { DConversation } from '~/common/state/store-chats';
