+++ src/modules/llms/openai/openai.vendor.ts
@@ -4,7 +4,7 @@ import { OpenAIIcon } from '~/common/components/icons/OpenAIIcon';
 
 import { DLLM } from '../store-llms';
 import { IModelVendor } from '../vendors/IModelVendor';
-import { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../llm.client';
+import { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../transports/chatGenerate';
 
 import { OpenAILLMOptions } from './OpenAILLMOptions';
 import { OpenAISourceSetup } from './OpenAISourceSetup';
