+++ src/modules/llms/vendors/IModelVendor.ts
@@ -1,7 +1,7 @@
 import type React from 'react';
 
 import type { DLLM, DModelSourceId } from '../store-llms';
-import type { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../llm.client';
+import { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../transports/chatGenerate';
 
 
 export type ModelVendorId = 'anthropic' | 'azure' | 'localai' | 'oobabooga' | 'openai' | 'openrouter';
